+++ signin/internal/web/web.go
@@ -2,14 +2,27 @@ package web
 
 import (
 	"encoding/json"
+	"html/template"
+	"log"
 	"net/http"
 
 	"github.com/weekendprojectapp/authful/server"
 	"github.com/weekendprojectapp/authful/signin/pkg/models"
 )
 
+type Student struct {
+	Name       string
+	College    string
+	RollNumber int
+}
+
 func DisplayLogin(w http.ResponseWriter, r *http.Request) {
-	server.HandleResponse(w, []byte{}, http.StatusOK)
+	parsedTemplate, _ := template.ParseFiles("template/login.html")
+	err := parsedTemplate.Execute(w, nil)
+	if err != nil {
+		log.Println("Error executing template :", err)
+		return
+	}
 }
 
 func AuthorizeUser(w http.ResponseWriter, r *http.Request) {
@@ -22,4 +35,18 @@ func AuthorizeUser(w http.ResponseWriter, r *http.Request) {
 
 func GetProfile(w http.ResponseWriter, r *http.Request) {
 	server.HandleResponse(w, []byte{}, http.StatusOK)
+}
+
+func Index(w http.ResponseWriter, r *http.Request) {
+	student := Student{
+		Name:       "GB",
+		College:    "GolangBlogs",
+		RollNumber: 1,
+	}
+	parsedTemplate, _ := template.ParseFiles("Template/index.html")
+	err := parsedTemplate.Execute(w, student)
+	if err != nil {
+		log.Println("Error executing template :", err)
+		return
+	}
 }
