+++ signin/template/login.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/resources/css/style.css">
+    <title>Go Template</title>
+</head>
+
+<body>
+    <span class="error">Login failed</span><br /><br />
+ 
+    <form method="post" id="loginForm" role="form" data-qa="login-form">
+        <h2>Login</h2>
+        Username:<br />
+        <input type="text" name="username" data-qa="login-username" /><br /><br />
+        Password:<br />
+        <input type="password" name="password" data-qa="login-password" /><br /><br />
+        <button id="btn-login" type="submit" data-qa="login-button">Log In</button>
+    </form>
+</body>
+</html>
